+++ src/Traits/Feed/Base/HasVariantInfo.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TimothyDC\LightspeedEcomProductFeed\Traits\Feed;
+namespace TimothyDC\LightspeedEcomProductFeed\Traits\Feed\Base;
 
 trait HasVariantInfo
 {
@@ -12,6 +12,10 @@ trait HasVariantInfo
     protected function generateVariantInfo(array $lightspeedData): void
     {
         foreach ($this->getVariants($lightspeedData) as $variant) {
+            if ($this->variantSkip($lightspeedData, $variant)) {
+                continue;
+            }
+
             $this->feed[$this->variantTreeMainNode][$this->variantTreeChildNode][] = $this->variantFields($lightspeedData, $variant);
         }
     }
@@ -44,5 +48,10 @@ trait HasVariantInfo
             'stock_buy_minimum' => $variant['stockBuyMinimum'],
             'stock_buy_maximum' => $variant['stockBuyMaximum'],
         ];
+    }
+
+    protected function variantSkip(array $lightspeedData, array $variant): bool
+    {
+        return false;
     }
 }