+++ src/Traits/Feed/HasCategoryTreeStructureFlat.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace TimothyDC\LightspeedEcomProductFeed\Traits\Feed;
 
+use TimothyDC\LightspeedEcomProductFeed\Traits\Feed\Base\HasCategoryInfo;
+
 /**
  * This trait will generate something like:
  *
@@ -33,7 +35,8 @@ trait HasCategoryTreeStructureFlat
     protected function generateCategoryInfo(array $lightspeedData): void
     {
         // only get a select set of category fields
-        $categories = collect($lightspeedData['categories']);
+        $categories = collect($lightspeedData['categories'])
+            ->filter(fn ($category) => $this->categorySkip($lightspeedData, $category) === false);
 
         // loop through each depth and add convert flat categories to a tree structure
         foreach ($categories->pluck('depth')->unique()->sort() as $depth) {