+++ tests/Feature/Feeds/SooqrFeedTest.php
@@ -251,31 +251,14 @@ class SooqrFeedTest extends TestCase
                     ],
                 ],
             ],
-            'filters' => [
+            'color' => [
                 'node' => [
-                    [
-                        'title' => ['_cdata' => 'Color'],
-                        'values' => [
-                            'node' => [
-                                ['_cdata' => 'Red'],
-                                ['_cdata' => 'Green'],
-                            ],
-                        ],
-                    ],
+                    ['_cdata' => 'Red'],
+                    ['_cdata' => 'Green'],
                 ],
             ],
-            'specifications' => [
-                'node' => [
-                    [
-                        'title' => ['_cdata' => 'Packages'],
-                        'value' => ['_cdata' => 5],
-                    ],
-                    [
-                        'title' => ['_cdata' => 'Airtight'],
-                        'value' => ['_cdata' => 'Yes'],
-                    ],
-                ],
-            ],
+            'packages' => [['_cdata' => 5]],
+            'airtight' => [['_cdata' => 'Yes']],
         ];
     }
 
@@ -312,27 +295,12 @@ class SooqrFeedTest extends TestCase
     <sub_sub_categories>
       <node><![CDATA[Category 1-1-1]]></node>
     </sub_sub_categories>
-    <filters>
-      <node>
-        <title><![CDATA[Color]]></title>
-        <values>
-          <node><![CDATA[Red]]></node>
-          <node><![CDATA[Green]]></node>
-        </values>
-      </node>
-    </filters>
-    <specifications>
-      <node>
-        <title><![CDATA[Packages]]></title>
-        <value>
-          <_cdata>5</_cdata>
-        </value>
-      </node>
-      <node>
-        <title><![CDATA[Airtight]]></title>
-        <value><![CDATA[Yes]]></value>
-      </node>
-    </specifications>
+    <color>
+      <node><![CDATA[Red]]></node>
+      <node><![CDATA[Green]]></node>
+    </color>
+    <packages><![CDATA[5]]></packages>
+    <airtight><![CDATA[Yes]]></airtight>
   </product>
 </products>
 ';